Finished Level Manager (mostly)
I've done as much as I can with the level manager.

Now I need to revamp the combat manager before coming back to finish one
of the big chunks left in it.

Then I need to work on the turn manager so I can have a game loop.

After that it's Input Manager, but Unity's works fine for now. I just
need to not rely on it since Inputs can't be changed at run time (which
means building my own).