Made CombatStats a struct + CombatManger changes
I updated the combats stats to be a structure instead of a class. This
makes it much easier for me to keep the values tied to rpgclass, as well
as initialize it properly.

I also got my combat manager working to the point where I can leave it
alone for now. It does all the math great but the animations aspect are
kind of a mess. I'll make the math bits cleaner when I come back to
flesh out the animation aspects.

Next on the list is creating content. Creating some new unit classes,
working on animations, more weapons, and some actual levels.