Added Combat Information Menu
There's now a combat info menu that pops up when the player selects a
unit for combat. It's missing the number of times a unit will attack,
but other than that has everything.

The new menu caused some bugs with the state machine. I'll have to go
back and fix them or else the player will get stuck in a 'combat loop'.