Made + Imported New Art
Made new art for mountain and fort tiles. Also added art for the
wanderer class. It didn't turn out how I wanted it to for wanderer (way
too big). But it gave me an idea for a new class. (Highwayman or Raider
or something. Big thug guy with a big swrd.)