Turn Manager
Created the turn manager and have it working with C# events. It keeps
track of the units that can be used, the turn count, and who is in
control of the turn. Then it triggers these events. There are separate
turn start and turn end events.

The next step is getting the objects to subscribe to these updates and
then do stuff.

Some leftover music stuff that I forgot to push is in here too.